PR notes
