+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Changelog of ai-cockpit.
 
 ### Features
 
+-  ab#1378: Added autonomy level information ([b22a0](https://github.com/starwit/ai-cockpit/commit/b22a0dfd25a1962) Anett Hübner)  
+-  added dialog for automomy level ([22660](https://github.com/starwit/ai-cockpit/commit/2266050b8efd97f) Anett Hübner)  
+
+## refs/tags/0.1.13 (2025-01-31)
+
+### Features
+
 -  add action execution ([8bdff](https://github.com/starwit/ai-cockpit/commit/8bdffce44a78c7d) Anett Hübner)  
 
 ## refs/tags/0.1.12 (2025-01-23)
