+++ application/src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.profiles.active=dev
 spring.banner.location=classpath:banner.txt
 server.servlet.context-path=/ai-cockpit
 rest.base-path=/api
-server.port=8081
+server.port=8080
 
 # actuator
 management.endpoints.web.base-path=/monitoring
